Use fixed width instead of height for main cover art on single page
Using a fixed height posed problems with narrow viewports (such as
on mobile devices). Set a fixed width and ensure that does not exceed
the usable width of the page, while allowing the height to be dynamic.

Also, adjust the height to match a whole number of lines of text
(for vertical rhythm).