Use fixed width as well as height for cover art in list
The centering and cropping for wide images wasn't actually behaving
as intended; it only seemed that way because of the particular
test image used.

The compromise now is that tall images will be cropped (not centered).

Also, adjust the height to match a whole number of lines of text
(for vertical rhythm).